pyproject.toml
[tool.poetry]
name = "rnb-backend-test"
version = "0.1.0"
description = "Rock'n'Block Backend Test"
authors = ["Timur Israpilov <very1fake.coder@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.13"
djangorestframework = "3.13.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
